separate VirtualBackend api definition from common utilities
