+++ virtualizarr/readers/kerchunk.py
@@ -3,7 +3,7 @@ from typing import Iterable, Mapping, Optional
 import ujson
 from xarray import Dataset, Index
 
-from virtualizarr.readers.common import VirtualBackend
+from virtualizarr.readers.api import VirtualBackend
 from virtualizarr.translators.kerchunk import dataset_from_kerchunk_refs
 from virtualizarr.types.kerchunk import (
     KerchunkStoreRefs,
