+++ virtualizarr/readers/netcdf3.py
@@ -3,11 +3,11 @@ from typing import Iterable, Mapping, Optional
 
 from xarray import Dataset, Index
 
-from virtualizarr.readers.common import (
-    VirtualBackend,
+from virtualizarr.readers.api import VirtualBackend
+#from virtualizarr.readers.common import (    
     # construct_virtual_dataset,
     # maybe_open_loadable_vars_and_indexes,
-)
+#)
 from virtualizarr.translators.kerchunk import (
     virtual_vars_and_metadata_from_kerchunk_refs,
 )
